+++ public/Public/homePage.html
@@ -4,9 +4,9 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-    <link rel="stylesheet" href="index.css" />
-    <link rel="stylesheet" href="styles.css">
-  
+    <link rel="stylesheet" href="public/css/index.css" type="text/css" />
+    <link rel="stylesheet" href="public/css/styles.css" type="text/css" />
+
     <title>Home Page</title>
 </head>
 <body>
@@ -16,9 +16,9 @@
       <div class="row navbar">
         <div class="col-10">
           <!-- TODO: Hide anchors until user has logged in -->
-          <img src="static/bow.png" alt="" id="logo">
-          <a href="/public" class="my-anchor">Home</a>
-          <a href="/public/recipes" class="my-anchor">Recipes</a>
+          <img src="public/img/bow.png" alt="" id="logo">
+          <a href="public" class="my-anchor">Home</a>
+          <a href="public/recipes" class="my-anchor">Recipes</a>
           <a href="" class="my-anchor">Best Sellers</a>
           <a href="" class="my-anchor">Stores</a>
           <a href="" class="my-anchor">Features</a>
@@ -37,7 +37,7 @@
   <div class="container-fluid backgroundColor">
     <div class="container">
       <div class="row">
-        <img src="images/foodBanner.jpg" alt="" height="350" width="1140">
+        <img src="img/foodBanner.jpg" alt="" height="350" width="1140">
         <div class="hunterGather">Hunter Gather</div>
       </div>
       <div class="row green-line"></div>
@@ -48,16 +48,16 @@
       </div>
       <div class="row">
         <div class="col-6 recipe-in-top">
-          <a href="recipe.html" style="color:black;">Grilled Cheese</a>
-          <img src="images/grilled-cheese.jpg" height="150px" width="150px">
+          <a href="" style="color:black;">Grilled Cheese</a>
+          <img src="uploads/grilled-cheese.jpg" height="150px" width="150px">
           <a href="" style="color:black;">Chicken Tikka Masala</a>
-          <img src="images/tikkaMasala.jpg" height="150px" width="150px">
+          <img src="img/tikkaMasala.jpg" height="150px" width="150px">
         </div>
         <div class="col-6 recipe-in-top">
           <a href="" style="color:black;">Pot Roast</a>
-          <img src="images/potroast.jpg" height="150px" width="150px">
+          <img src="img/potroast.jpg" height="150px" width="150px">
           <a href="" style="color:black;">Nachos</a>
-          <img src="images/nachos.jpg" height="150px" width="150px">
+          <img src="img/nachos.jpg" height="150px" width="150px">
         </div>
       </div>
       <div class="row green-line"></div>
@@ -69,19 +69,19 @@
       <div class="row">
         <div class="col-6 recipe-in-top">
           <a href="" style="color:black;">John Smith</a>
-          <img src="images/chef1.jpg" height="150px" width="150px">
+          <img src="img/chef1.jpg" height="150px" width="150px">
           <a href="" style="color:black;">Eva</a>
-          <img src="images/chef2.jpg" height="150px" width="150px">
+          <img src="img/chef2.jpg" height="150px" width="150px">
         </div>
         <div class="col-6 recipe-in-top">
           <a href="" style="color:black;">Christian</a>
-          <img src="images/chef3.jpg" height="150px" width="150px">
+          <img src="img/chef3.jpg" height="150px" width="150px">
           <a href="" style="color:black;">Alec</a>
-          <img src="images/chef4.jpg" height="150px" width="150px">
+          <img src="img/chef4.jpg" height="150px" width="150px">
         </div>
       </div>
 
-    </div>  
+    </div>
   </div>
 
   <footer>
@@ -100,4 +100,4 @@
 
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
 </body>
-</html>  
+</html>
