+++ public/Public/recipeList.html
@@ -4,9 +4,9 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-    <link rel="stylesheet" href="index.css" />
-    <link rel="stylesheet" href="styles.css">
-  
+    <link rel="stylesheet" href="css/index.css" type="text/css" />
+    <link rel="stylesheet" href="css/styles.css" type="text/css" />
+
     <title>Recipes</title>
 </head>
 <body>
@@ -16,7 +16,7 @@
       <div class="row navbar">
         <div class="col-10">
           <!-- TODO: Hide anchors until user has logged in -->
-          <img src="static/bow.png" alt="" id="logo">
+          <img src="img/bow.png" alt="" id="logo">
           <a href="/public" class="my-anchor">Home</a>
           <a href="/public/recipes" class="my-anchor">Recipes</a>
           <a href="" class="my-anchor">Best Sellers</a>
@@ -55,7 +55,7 @@
       <div class="row" style="margin-bottom:30px;">
         <div class="col-6 recipe-in-top">
           <a href="recipe.html" style="color:black;">Grilled Cheese</a>
-          <img src="images/grilled-cheese.jpg" height="150px" width="150px">
+          <img src="/uploads/grilled-cheese.jpg" height="150px" width="150px">
           <a href="" style="color:black;">Chicken Tikka Masala</a>
           <img src="images/tikkaMasala.jpg" height="150px" width="150px">
           <a href="" style="color:black;">Meatloaf</a>
@@ -84,7 +84,7 @@
       </div>
       </div>
 
-    </div>  
+    </div>
   </div>
 
   <footer>
@@ -103,4 +103,4 @@
 
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
 </body>
-</html>  
+</html>
